task2/requirements.txt
psycopg2-binary==2.9.9
requests==2.31.0
Flask==3.0.0
hazelcast-python-client==5.4.0
